fix 12 nov 2019

+++ Texture.cpp
@@ -7,7 +7,7 @@ TGAHeader tgaheader;									// TGA header
 TGA tga;												// TGA image data
 
 GLubyte uTGAcompare[12] = {0,0,2,0,0,0,0,0,0,0,0,0};	// Uncompressed TGA Header
-GLubyte cTGAcompare[12] = {0,0,10,0,0,0,0,0,0,0,0,0};	// Compressed TGA Header
+GLubyte cTGAcompare[12] = {0,0,10,0,0,0,0,0,0,0,0,0};	// Compressed TGA Header	
 
 bool LoadTGA(Texture * texture, char * filename)				// Load a TGA file
 {
@@ -22,7 +22,7 @@ bool LoadTGA(Texture * texture, char * filename)				// Load a TGA file
 
 	if(fread(&tgaheader, sizeof(TGAHeader), 1, fTGA) == 0)					// Attempt to read 12 byte header from file
 	{
-		//MessageBox(NULL, "Could not read file header", "ERROR", MB_OK);		// If it fails, display an error message
+		//MessageBox(NULL, "Could not read file header", "ERROR", MB_OK);		// If it fails, display an error message 
 		if(fTGA != NULL)													// Check to seeiffile is still open
 		{
 			fclose(fTGA);													// If it is, close it
@@ -30,7 +30,7 @@ bool LoadTGA(Texture * texture, char * filename)				// Load a TGA file
 		return false;														// Exit function
 	}
 
-	if(memcmp(uTGAcompare, &tgaheader, sizeof(tgaheader)) == 0)				// See if header matches the predefined header of
+	if(memcmp(uTGAcompare, &tgaheader, sizeof(tgaheader)) == 0)				// See if header matches the predefined header of 
 	{																		// an Uncompressed TGA image
 		LoadUncompressedTGA(texture, filename, fTGA);						// If so, jump to Uncompressed TGA loading code
 	}
@@ -47,22 +47,22 @@ bool LoadTGA(Texture * texture, char * filename)				// Load a TGA file
 	return true;															// All went well, continue on
 }
 
-bool LoadUncompressedTGA(Texture * texture, char * filename, FILE * fTGA)	// Load an uncompressed TGA (note, much of this code is based on NeHe's
+bool LoadUncompressedTGA(Texture * texture, char * filename, FILE * fTGA)	// Load an uncompressed TGA (note, much of this code is based on NeHe's 
 {																			// TGA Loading code nehe.gamedev.net)
 	if(fread(tga.header, sizeof(tga.header), 1, fTGA) == 0)					// Read TGA header
-	{
+	{										
 		//MessageBox(NULL, "Could not read info header", "ERROR", MB_OK);		// Display error
 		if(fTGA != NULL)													// if file is still open
 		{
 			fclose(fTGA);													// Close it
 		}
 		return false;														// Return failular
-	}
+	}	
 
 	texture->width  = tga.header[1] * 256 + tga.header[0];					// Determine The TGA Width	(highbyte*256+lowbyte)
 	texture->height = tga.header[3] * 256 + tga.header[2];					// Determine The TGA Height	(highbyte*256+lowbyte)
 	texture->bpp	= tga.header[4];										// Determine the bits per pixel
-	tga.Width		= texture->width;										// Copy width into local structure
+	tga.Width		= texture->width;										// Copy width into local structure						
 	tga.Height		= texture->height;										// Copy height into local structure
 	tga.Bpp			= texture->bpp;											// Copy BPP into local structure
 
@@ -104,7 +104,7 @@ bool LoadUncompressedTGA(Texture * texture, char * filename, FILE * fTGA)	// Loa
 	}
 
 	// Byte Swapping Optimized By Steve Thomas
-	for(GLuint cswap = 0; (int)cswap < (int)tga.imageSize; cswap += tga.bytesPerPixel)
+	for(GLuint cswap = 0; cswap < (int)tga.imageSize; cswap += tga.bytesPerPixel)
 	{
 		texture->imageData[cswap] ^= texture->imageData[cswap+2] ^=
 		texture->imageData[cswap] ^= texture->imageData[cswap+2];
@@ -115,7 +115,7 @@ bool LoadUncompressedTGA(Texture * texture, char * filename, FILE * fTGA)	// Loa
 }
 
 bool LoadCompressedTGA(Texture * texture, char * filename, FILE * fTGA)		// Load COMPRESSED TGAs
-{
+{ 
 	if(fread(tga.header, sizeof(tga.header), 1, fTGA) == 0)					// Attempt to read header
 	{
 		//MessageBox(NULL, "Could not read info header", "ERROR", MB_OK);		// Display Error
@@ -161,7 +161,7 @@ bool LoadCompressedTGA(Texture * texture, char * filename, FILE * fTGA)		// Load
 
 	GLuint pixelcount	= tga.Height * tga.Width;							// Nuber of pixels in the image
 	GLuint currentpixel	= 0;												// Current pixel being read
-	GLuint currentbyte	= 0;												// Current byte
+	GLuint currentbyte	= 0;												// Current byte 
 	GLubyte * colorbuffer = (GLubyte *)malloc(tga.bytesPerPixel);			// Storage for 1 pixel
 
 	do
@@ -209,7 +209,7 @@ bool LoadCompressedTGA(Texture * texture, char * filename, FILE * fTGA)		// Load
 					return false;														// Return failed
 				}
 																						// write to memory
-				texture->imageData[currentbyte		] = colorbuffer[2];				    // Flip R and B vcolor values around in the process
+				texture->imageData[currentbyte		] = colorbuffer[2];				    // Flip R and B vcolor values around in the process 
 				texture->imageData[currentbyte + 1	] = colorbuffer[1];
 				texture->imageData[currentbyte + 2	] = colorbuffer[0];
 
@@ -228,7 +228,7 @@ bool LoadCompressedTGA(Texture * texture, char * filename, FILE * fTGA)		// Load
 					if(fTGA != NULL)													// If there is a file open
 					{
 						fclose(fTGA);													// Close file
-					}
+					}	
 
 					if(colorbuffer != NULL)												// If there is data in colorbuffer
 					{
@@ -248,7 +248,7 @@ bool LoadCompressedTGA(Texture * texture, char * filename, FILE * fTGA)		// Load
 		{
 			chunkheader -= 127;															// Subteact 127 to get rid of the ID bit
 			if(fread(colorbuffer, 1, tga.bytesPerPixel, fTGA) != tga.bytesPerPixel)		// Attempt to read following color values
-			{
+			{	
 				//MessageBox(NULL, "Could not read from file", "ERROR", MB_OK);			// If attempt fails.. Display error (again)
 
 				if(fTGA != NULL)														// If thereis a file open
@@ -269,7 +269,7 @@ bool LoadCompressedTGA(Texture * texture, char * filename, FILE * fTGA)		// Load
 				return false;															// return failed
 			}
 
-			for(short counter = 0; counter < chunkheader; counter++)					// copy the color into the image data as many times as dictated
+			for(short counter = 0; counter < chunkheader; counter++)					// copy the color into the image data as many times as dictated 
 			{																			// by the header
 				texture->imageData[currentbyte		] = colorbuffer[2];					// switch R and B bytes areound while copying
 				texture->imageData[currentbyte + 1	] = colorbuffer[1];
@@ -290,7 +290,7 @@ bool LoadCompressedTGA(Texture * texture, char * filename, FILE * fTGA)		// Load
 					if(fTGA != NULL)													// If there is a file open
 					{
 						fclose(fTGA);													// Close file
-					}
+					}	
 
 					if(colorbuffer != NULL)												// If there is data in colorbuffer
 					{
